Create Marathon
Define Java interface IMarathon with the following methods:

Athlete bestTime() - that returns the winner of the marathon

int AthletesFrom(String s) - that returns the number of athletes from s country.

Define Java class Athlete. For every athlete the following information is stored: name (String), gender (String), age (int), time score in seconds (double) and country of origin (String). For this class implement:

constructors (default and with parameters)
set and get methods
the method toString() in the following format: name / age / country of origin / time score (all of the parameters should be split with one space)
Define Java class Marathon that implements the interface IMarathon. For every marathon the following information is stored: host country (String), year (int), an array of Athletes Athlete[]. For this class implement:

constructors (default and with parameters)
set and get methods
the method toString() in the following format: host country / year / attendees of the marathon (all of the parameters should be split with one space)
bestTime()
AthletesFrom(String s)