Create Average
For a given array with N (1≤N≤50) integers, find the element which is the closest to the average. If there are two elements with the same distance to the average, return the smaller. For example for the array 1, 2, 3, 4, 5 the average is (1 + 2 + 3 + 4 + 5) / 5 = 15 / 5 = 3, which means that the element which is the closest to the average is 3.

For the array 1, 2, 3, 4, 5, 6 the average is 3.5 and both 3 and 4 are equally distant to the average. The correct answer is the smaller value which is 3.

There can be duplicates in the array.

The first number in the input is the number of elements in the array N, then the elements of the array follow in each line.

Class (Java): Array

**Note:** Create data structure array and use it in this problem.