Create Special join lists
Given two single linked lists with integer nodes, join them in a resulting single linked list in a way that alternatively adds at first the first two nodes from the first list, then the first two from the second, then next two from the first list, and then the second two from the second list, etc. The nodes that remain should be added at the end, at first from the first list, then the remaining from the second list.

In the first line from the input the numbers from the first list are given, and in the second line the numbers from the second list. The numbers from the resulting list should be printed at the output.

Note: Create data structure single linked list and use it in this problem.