Create Arithmetic Expression
You are given an arithmetic expression. The arithmetic expression is of type (A+B) or (A-B) where А and B can also be arithmetic expressions or a number 0-9. You should evaluate the given expression.

Input: The expression to be evaluated in one line

Output: Value of the expression

Class name (Java): ArithmeticExpression