Create Sum of absolute differences (Exam problem)
You are given an array of N natural numbers and one number K. Let the numbers be marked from a0 to aN−1. We define the sum of absolute differences as: abs(a1−a0)+abs(a2−a1)+…+abs(aN−1−aN−2). Select exactly K numbers from the array, such that, when put together, the sum of absolute differences is maximal. Print this sum.

Input: In the first line are the two numbers N (1≤N≤100) and K (1≤K≤100, K≤N). In the second line there are N positive natural numbers, each of them smaller than 1,000.

Output: Print the required maximum sum of absolute differences.

NOTE: The order of the numbers should not be changed from the original array.

Name of the class (Java): SumOfAbsoluteDifferences.

Partial solution: The problem is considered partially solved with 5 correct test cases.