Create Counter of the Ministry of Internal Affairs
Every morning people come to the Ministry of Internal Affairs to apply for one or more documents.
The document can be one of the following:
1. ID card
2. Passport
3. Driving license
People waiting for ID card are served first when the counter opens, then those for passport and finally those waiting for driving license. Every person waits in the queue for the appropriate document that he/she applies for (i.e. either in the queue for ID cards or in the queue for passports in the queue or for driving licenses). If a person has to apply for more documents, first he/she applies for ID card, then for passport and finally for driving license. So if a person applies for ID card and driving license he/she first waits in the queue for ID cards and when he/she finishes, he/she goes at the end of queue for driving licenses.

The first input line indicates how many people came to MIA. Then, four lines are entered for every person (first name and surname are entered in the first line, and the three other lines are for the specifying the appropriate document (identity card, passport and driving license) that a person applies for. 1 means that the person applies for the document and 0 means that he/she does not apply.

For example:

Aleksandar Aleksandrovski
1
0
1
means that Alexander Aleksandrovski will apply for ID card and driving license, but not for passport.
Output should print the names of people in order they finish with applying.

Class name (Java): MVR