Create Odd even sorting
You are given an array of N integers. The odd numbers should be put in the beginning of the array in ascending order, and the even numbers should be put at the end of the array in descending order.

In the first line of the input you will find a single integer N, representing the number of integers in the array. In the second line of the input you will the elements of the array with a single whitespace between them.

Class name: OddEvenSort