Create Shaker (cocktail) sort
You are given array of N integers. This array should be sorted using the so called shaker(cocktail) sorting. This sorting algorithm is a variation of bubble sort. Every iteration of the algorithm is consisted of two parts. In the first part you are supposed to put the smallest element at the beginning of the array and in the second part you are supposed to put the biggest element at the end. You continue this process until the array is fully sorted.

In the first line of the input you will find a single integer N, representing the number of integers in the array. In the second line of the input you will find the elements of the array with a single whitespace between them.

The output should contain the state of the array after every iteration.

Class name: Shaker Sort