Create Bubble sort on a linked list
You are given a doubly linked list with N nodes, where each node contains a single integer. You are supposed to sort this list using the bubble sort algorithm.

In the first line of the input you will find a single integer N, representing the number of nodes in the list. In the second line of the input you will find the elements of the list with a single whitespace between them.

The output should contain the final state of the array after the sorting is done.

Note: You must use the data structure you've been given. You are not allowed to use arrays and similar data structures.

Name of the class: BubbleSortDLL