Create Translator
You are supposed to make an automated translator, that translates words from English to Macedonian.

Input: In the first line you are given a single integer N. In the following N lines you are given two words separated with a single white space. The first one is on Macedonian and the second is on English. Use this pairs of words to create, a dictionary. After this you are given a single English word in every line, until the word "КРАЈ" is read.

Output: On the standard output you should print the Macedonian translation of the English words, using the dictionary you previously created. If you can't find a word in the dictionary print "/"

Note. Use open buckets hash table. You are supposed to define the hash table and write the hash function by yourself.

Class name: Preveduvac