Create Black Friday
There is a Black Friday event in a local store with discounts up to 80%. The store were expecting a lot of buyers, so they decided to develop an ordering algorithm for entering and leaving the store, so that there would be no problems. The buyers enter the store according to the time of arrival. You are supposed to find the maximum number of buyers that can be present at the same time in the store. The store is open til 23:59. Use a heap structure (priority queue) to solve this problem.

In the first line of the input you are given the number of buyers N. In each of the following lines you are given the time of arrival for the i-th buyer and the minutes that he plan to spend in the store.

On the output print a single integer representing the maximum number of buyers present at the store, at the same time.

Name of class: BlackFriday