Create Binary tree
Write a recursive function that will find the height(V) of a given element inside a binary tree (all of the leaves have a height of 1). After that print the number of elements that can be found at depth V (the root has a depth of 0).

In the first line of the input you will find, a single integer N representing the number of nodes. In the following N lines you are given a single integer X-i representing the element of the i-th node. The last line contains the element you are supposed to search for in the tree.

The output should contain 2 lines. In the first line print the height(V) of the element you read in the last line. In the second line print the number of elements that can be found at depth V.

Class name: BinarnoDrvo