Create Maze traversal
It is given a perfect maze (in the same form as presented on the aud. exercises - input of characters). Your task is to create a graph (unoriented and unweighted) by using the adjacency list from the given input and to print the path from the initial vertex (vertex denoted by S) to the final vertex (vertex denoted by E).

Print the path as follows: print the coordinates of the unfolded vertices (i.e. print the positions from the input).

Class name: Lavirint