Create University Subjects
On one university there are mandatory and elective subjects. For each subject you are given, a list of subjects that need to be passed in order to be eligible for that subject is given. Find the next available subject for you, if you know what is the last subject that you have passed.

Input: In the first line you are given a single integer N representing the number of elements. In the following N lines you are given the IDs of the subjects. In the following line you are given a single integer M representing the number of dependences between the subjects. In the following M lines you are given a list of IDs separated with a single white space. Starting from the second ID to the last ID you have the subjects that you need to pass in order to be able to take the subjected defined with the first ID in this line. In the last line you are given the last subject that you have passed.

Output: The output should contain a single ID representing the next available subject.