Create Hierarchy
One organisation has a single president and multiple regular members. If you know the collaboration degree (1 being perfect, 10 being worst) between the members choose the best hierarchy for the organisation. By best it means that the sum of all collaboration degrees between the members should be minimum.

Input:

In the first line of the input you are given a single integer N, representing the number of members the organisation has. In the second line of the input you are given a single integer M, representing the number of collaborations between members from the organisation. In the following M lines you are given the collaborations like following: ID of a member, name of that member, ID of a member, name of that member, degree of collaboration between the two members given previously. In the last line you are given the name of the president.

Output:

After choosing the best hierarchy print the sum of all collaboration degrees for that hierarchy.