Create Cities
You are given a map with one way roads between N cities in Macedonia. For each road you are given the length of the road, the starting city and the destination city. Find the minimal distance from one city to another and back, and print the roads you are going to take.

The map with roads is given as a directed weighted graph.

Input: In the first line you are given the number of cities N. In the second line you are given the number of roads M. In the following M lines you are given the roads as following: the number of the starting city, the name of the starting city, the number of the destination city, the name of the destination city, length of the road. In the last 2 lines you are given the names of the starting and the destination city.

Output: In the first line of the output you are supposed to print the route which you are supposed to take starting from the starting city and ending at the destination city. In the second line you are supposed to print the route which you are supposed to take starting from the destination city and ending at the starting city. In the last line of the output you are supposed to print a single integer representing the total length of the routes.