Create Kumanovo Dialect
You are given a dictionary of words in the kumanovo dialect, and how they are written in the macedonian language. Then, you are given a text written in the kumanovo dialect. You need to swap all occurences of a dialect word with the corresponding standard macedonian word.

Note: You should avoid the punctuation marks, dot (.) , comma (,), exclamation point (!) and question mark (?). The swapped words should keep the case they were originally written in (case sensitive).

Name of the class(Java): KumanovskiDijalekt.