Create Frequent Substring
For a given string, find the most frequent substring. If two substrings have the same frequency, then print the longer one, and if they are the same, print the one higher lexicographically.

Example: For the string "abc" substrings are "a", "b", "c", "ab", "bc", "abc". All have the same frequency, but "abc" is the longest.

Name of the class (Java): MostFrequentSubstring.