Create Spell Checking
Write a program to check the spelling of a text. You are given a dictionary of english words, then a text that should be spell-checked. As a result, all incorrectly spelled words, or words not in the dictionary should be printed.

Input: First you input the number N, then N lines of english words, then the text to be checked.

Output: Print a list of incorrectly spelled words (each on its own line). If all words are correct, print: Bravo.

Note: Avoid punctuation marks such as, dot (.), comma (,), exclamation point (!) and question makr (?). Words are case insensitive, i.e. words with first capital letter in the text should be considered correct even if they don't start with capital in the dictionary. Work with open bucket hash table (OBHT). You are left to decide the number of buckets in the table.

Class name (Java): Speluvanje.