Create Total number of Roads
Print the total number of paths with length N, that start from a fixed vertex V, in an undirected, unweigthed graph.

Input: The first and second line contain the number of vertices and edges, then follow the edges of the graph. After than, the starting vertex V is given, and in the last line is the number N.

Output: Print the total number of paths with length N. Vertices may repeat.

In the sample test case, there are two paths 0 1 2 and 0 1 0, Which start from 0, and have a length of 2.

Class name (Java): PathCount.