Create Components of Connection
For a given vertex, print the vertices of the same component of a given graph. Vertices belong to the same component if there is a path between them. First you are given N, the number of vertices, then for each vertex, the number of vertices it is connected with, and the ids of the vertices. Lastly, the id of the required vertex is given.

Class Name (Java): KomponentiSvrzanost.